refactor mapping
