+++ mapping_models/src/.gitignore
@@ -1,0 +1,1 @@
+checkpoints
